Generate Dataset ...

Generate Dataset ...

Create an empty assumptions data.frame for generate_diabetes_rescue

Calculate true summary statistics for scenarios with delayed treatment effect

Usage

generate_diabetes_rescue(condition, fixed_objects = NULL)

assumptions_diabetes_rescue(print = interactive())

true_summary_statistics_diabetes_rescue(
  Design,
  cutoff_stats = 10,
  fixed_objects = NULL
)

Arguments

condition

condition row of Design dataset

fixed_objects

fixed objects not used for now

print

print code to generate parameter set?

Design

Design data.frame for x

cutoff_stats

Cutoff time for rmst and average hazard ratios

Value

For generate_diabetes_rescue: A data set with the columns id, trt (1=treatment, 0=control), evt (event, currently TRUE for all observations)

For assumptions_diabetes_rescue: a design tibble with default values invisibly

For true_summary_statistics_x: the design data.frame passed as argument with the additional columns:

rmst_trt rmst in the treatment group

median_surv_trt median survival in the treatment group

rmst_ctrl rmst in the control group

median_surv_ctrl median survial in the control group

gAHR geometric average hazard ratio

AHR average hazard ratio

Details

Condition has to contain the following columns:

eff the effect size

...

assumptions_diabetes_rescue generates a default design data.frame for use with generate_diabetes_rescue If print is TRUE code to produce the template is also printed for copying, pasting and editing by the user. (This is the default when run in an interactive session.)

Functions

generate_diabetes_rescue(): simulates a dataset with ...

assumptions_diabetes_rescue(): generate default design tibble

true_summary_statistics_diabetes_rescue(): calculate true summary statistics for ...

Examples

generate_diabetes_rescue(assumptions_diabetes_rescue())
#>    id trt      age        y0        y1        y2        y3        y4        y5
#> 1   1   0 48.06359  7.621971  9.783763  7.250056  7.181384  9.208099  6.742206
#> 2   2   0 52.48277  7.330467  7.556490  9.389615  8.741800  8.654645  6.424021
#> 3   3   0 74.55841  8.719016  7.309590  6.121063  8.099083  8.905687  8.491801
#> 4   4   0 51.71396  8.115348  6.268951  6.666360  8.820176  6.907154  7.816669
#> 5   5   1 62.89774  8.386631  7.303582  6.800106  8.420043  6.508635  8.665624
#> 6   6   1 55.19947  7.076062  7.773561  7.498920  8.009110  9.347322  8.842642
#> 7   7   0 53.95171  7.832771  8.517676  6.729971  7.833084  8.779120  8.907272
#> 8   8   0 74.60110  8.064544  7.742425  7.567800  6.612016  7.523870  7.768099
#> 9   9   0 61.49679  8.197958  6.815214  7.997743  8.743323  8.780388  9.085343
#> 10 10   0 45.66679  8.904269  8.135913  6.853701  9.194477  7.494278  8.091175
#> 11 11   0 59.89697  7.217621  7.356878  8.668275  7.555700  8.354292  8.572390
#> 12 12   0 57.87764  9.367707  7.396700  7.831641  7.388691  7.651837  8.727165
#> 13 13   1 50.93660  7.194048  8.247367  7.919739  9.985562  7.775149  9.070787
#> 14 14   1 38.97848  8.537885  7.912680  7.577525  7.221450  7.897601  8.205108
#> 15 15   0 78.93360  7.904865  5.510243  8.178105  8.367316  9.456133  8.793946
#> 16 16   0 50.31874  7.256091  7.931715  7.520683  8.310270  8.911876  6.287243
#> 17 17   0 58.97397  8.476844  7.227167  8.279065  6.844656  5.635828  6.622584
#> 18 18   0 62.39960  9.170056  6.612926  9.136990  8.756605  9.103737  7.832223
#> 19 19   1 60.60899  8.289090  8.953505  6.987697  7.047270  6.787627  7.764857
#> 20 20   0 38.22424  8.225340  7.364333  6.966010  8.689551  7.625982  8.309126
#> 21 21   0 58.82140  8.736068  6.800105  9.066629  8.532985  8.880900  7.982769
#> 22 22   1 61.12295  7.642472  7.446426  8.114754  8.638119  7.634030  9.454541
#> 23 23   1 60.07886 10.167187 10.594017  7.449057  8.664384  9.606062  8.427821
#> 24 24   1 78.77744  8.550016  8.072109  8.969515  7.724505  8.726079  8.130509
#> 25 25   1 81.58757  8.136058  6.237571  7.984784  8.147515  9.833368  7.134215
#> 26 26   0 67.09715  7.979890  7.903521  8.290138  8.342063  7.867565  8.623341
#> 27 27   1 67.66983  8.121621  7.216625  7.363091  8.130481  8.090955  8.652636
#> 28 28   1 56.91789  8.063919  7.141225  9.012267  7.356380  8.860415  8.381893
#> 29 29   1 70.12002  6.434356  8.092828  8.763823  8.877067  9.252759  9.113180
#> 30 30   1 50.80948  7.038723  8.447849  7.795210  9.387052  8.012660  8.175621
#> 31 31   0 65.63380  9.404079  9.391216  9.088592  8.674468  9.141321  9.426951
#> 32 32   1 63.22483  7.479456  8.877057  8.384569  7.242288  7.873073  6.565799
#> 33 33   1 63.66674  8.037735  6.284609  7.083906  7.847455  7.122281 10.808830
#> 34 34   1 71.29835  5.905029  8.275601  9.267821  9.283841  9.189311  7.451095
#> 35 35   0 50.58502  8.521976  6.121689  6.097372  8.641233  8.937827  7.387935
#> 36 36   1 62.17838  7.708122  6.909977  6.947343 10.152152  7.213733  8.604581
#> 37 37   0 74.15412  7.084079  8.081078  9.253792  8.408344  8.714107  7.421160
#> 38 38   0 56.16267  9.182088  8.216394  8.303312  9.128188  6.819425  8.511023
#> 39 39   1 58.25914  8.954146  7.561529  6.785928  6.971361  7.071601  7.735700
#> 40 40   0 57.78255  7.584205  9.236784  9.288949  8.758560  8.070380  8.431054
#> 41 41   1 49.90471  8.017758  7.349637  9.727069  9.709983 10.049902  8.593827
#> 42 42   1 64.80725  8.056790  8.133235 10.358985  9.411038  7.318906        NA
#> 43 43   0 76.04407  8.440355  9.326634  8.260264  8.962381  7.685255  7.758111
#> 44 44   0 44.84975  8.245949  8.331602  7.126626  8.151049 10.461015  8.453216
#> 45 45   1 45.83976  8.660769  7.281512  8.632428  8.648040  6.701936  6.493058
#> 46 46   1 68.76777  9.505859  7.075770  8.407356  8.582040  6.890772  9.838072
#> 47 47   1 66.24132  8.860613  7.991352  7.496191  7.810289  7.961734  8.824281
#> 48 48   0 81.12277  6.623169  8.887989  8.681057  9.244477  9.664741  7.841166
#> 49 49   0 56.43876  7.803681  8.941726  9.131969  9.532802  6.453384  5.371464
#> 50 50   0 49.35536  8.724691  5.666056  8.021732  7.661007  7.893766  8.792690
#> 51 51   1 70.77117  7.850412  8.532445  8.985858  8.016142  9.638255  8.063670
#> 52 52   0 71.81576  9.138668  8.199708  6.651976  6.729452  5.899641  7.949718
#> 53 53   0 61.98392  8.398259  6.543619  7.640872  7.923768  6.073133  7.094853
#> 54 54   0 55.99595  6.223041  7.599663  6.023968 10.833257  9.623660  8.171476
#> 55 55   1 66.16154  7.144587  7.962079  7.862161  8.074329  7.057440  6.428138
#> 56 56   0 79.74157  7.481654  7.321283  7.124956  8.065144  8.206771  7.417855
#> 57 57   0 78.84662  8.355433  7.021643  8.226013  7.936840  6.600198  9.136358
#> 58 58   1 44.11379  6.448440  9.047620  9.547904  8.462262 10.340098  8.683135
#> 59 59   1 54.60077  6.557243  9.758160  7.762294  8.362001  9.126338  8.276638
#> 60 60   0 48.30539  6.792141  8.374678  7.559555  9.424527  8.216371  7.838623
#> 61 61   0 65.59106  7.426099  7.308181  6.170876  7.795450  7.720999  7.821323
#> 62 62   1 41.80653  7.026506  7.807265  7.644342  7.270173  8.759340 10.642656
#> 63 63   1 63.93344  8.743620  8.072392  9.526998  8.369075  7.706792  9.034248
#> 64 64   0 60.42134  9.402474  7.296365  9.371048  7.249368  9.875971  8.787795
#> 65 65   0 71.79664  8.255815  7.811121  8.447183  8.098361  6.500134  8.499969
#> 66 66   1 57.43079  7.975794  8.897777  7.381312  8.716531  6.649419  7.272569
#> 67 67   0 49.43664  7.409062  7.301227  8.780164  8.070514  6.309409  6.942121
#> 68 68   1 61.98777  7.337640  7.725262  8.730949  9.175264  7.045479  9.125611
#> 69 69   1 66.50534  8.105373  7.950002  7.375202  8.854416  9.555714  9.233144
#> 70 70   0 63.43913  8.326423  6.557071  7.233534  8.519257  6.915478  7.810476
#> 71 71   1 74.77532  6.746634  7.120126  8.239679  8.076095  8.896609  9.372841
#> 72 72   1 60.72026  7.615962  7.627210  7.612305  6.424787  7.198180  7.826772
#> 73 73   0 81.26445  7.605813  8.530029  8.828396  8.239377  8.519065  7.214605
#> 74 74   0 45.23803  6.336589  9.973652  7.299028  8.557534  7.773213  7.574103
#> 75 75   0 64.07888  8.670923  8.046657 10.399917 10.632896  9.275091  6.827316
#> 76 76   1 73.93978  8.959055  8.244328  7.389479  9.601737  9.164867  9.291154
#> 77 77   1 63.60278  7.752137  6.714639  8.115389  8.629738  8.451579  6.587086
#> 78 78   0 66.54550  6.862969  7.733515  7.132341  8.176406  7.303774  8.584373
#> 79 79   1 70.52155  7.170181  9.589645  8.144596  8.711032  7.432598  6.241188
#> 80 80   0 40.20445  9.674790  7.081180  9.387024  9.323670  7.716697  8.917671
#> 81 81   1 72.08386  8.316726  9.302793  9.101148  7.460163  7.880222  8.801446
#> 82 82   0 58.30720  6.697328  8.658285  9.046509  7.732532  8.007674  6.398318
#> 83 83   0 62.95030  8.548811  7.477613  7.862642  7.284206  7.461346        NA
#> 84 84   1 72.66341  9.166684  7.779340  7.776175  7.833491  7.340039  8.068210
#> 85 85   1 48.64657  7.715974  8.371902  8.566233  5.813414  8.928772  8.926548
#> 86 86   1 48.68946  6.885239  9.668641  7.494472  9.759197  6.797889  7.507470
#> 87 87   1 61.09993  7.176212  8.424149  8.065348  8.517419  7.781054  8.053162
#> 88 88   1 68.52905  7.806550  7.929113  8.416170  8.804845  9.075730  8.622020
#> 89 89   0 57.65662  8.647418  8.902971  8.399813  8.123403  8.933780  7.840813
#> 90 90   1 80.86689  6.806459  7.958373  7.426963  6.560000  8.703359        NA
#> 91 91   1 58.89081  9.341974  8.271643  9.078184 10.440606  8.499812  8.470943
#> 92 92   0 46.07153  7.701162  9.439329  7.409346 10.025054 10.022755  6.917494
#> 93 93   1 48.57709  7.020078  6.671389  9.581911  7.772558  7.517580  7.595409
#> 94 94   0 77.04609  9.747150  7.924687  8.136925  8.411411  8.403987  8.198134
#> 95 95   0 59.19926  8.032001  8.202420  8.121205  8.624462  7.921235  7.882131
#> 96 96   0 55.62719  8.881790  6.229466  7.099368  7.610570  6.252376  8.394737
#> 97 97   0 58.80785  9.414225  8.264254  6.609169  7.753331  7.609969  8.949394
#> 98 98   1 67.86463  7.495641  7.750315  7.811002 10.571497  9.906951  8.113250
#>           y6        y7        y8        y9       y10       y11       y12
#> 1   7.101093  8.963798  8.733264  9.502472  7.869084 10.068885  7.943638
#> 2   9.104073  8.468593  8.861450  8.137892  9.091759  7.388428  5.481973
#> 3   7.194851  8.619364  9.212969  7.312154  7.586688  6.917489  7.611467
#> 4   8.684076  7.936162  7.008779  7.302953  9.884082  8.753896  8.222386
#> 5   8.359609  9.904436  8.572070  9.841307  8.585521 10.531239  8.120187
#> 6   7.604617  7.458033  7.436107  6.157669  6.981015  7.893403  7.834384
#> 7   9.086660  9.711084  7.594063  8.822433  9.063636 10.244833  6.320129
#> 8   6.960278  8.415164  7.724905  9.431954  7.364198  8.745473  8.381997
#> 9   9.256930  6.892551 10.805140  8.012105  7.764312  8.008195  7.330560
#> 10  8.866449  8.944059  8.999797  7.846835  8.620063  5.810895  7.762909
#> 11  8.849891  8.760716  7.549992  8.409569  9.337753 10.044264  8.167966
#> 12        NA        NA        NA        NA        NA        NA        NA
#> 13  8.390558  6.882019 10.071994  6.041213  7.697552  8.064799  8.113021
#> 14  9.769577  9.363317 10.969392  8.366449  9.309663  8.978722  9.015740
#> 15  9.201168  7.902312  7.529961  9.570216  7.727587  8.008015  7.503770
#> 16  8.480633  8.914247  8.743975 10.343564  7.901276  7.338349  7.540369
#> 17  7.702066  7.214029  8.057501  9.537767  7.363333  8.912596  6.846893
#> 18  8.614450  8.866762  9.018884  9.743228  8.074365 10.242145  8.086147
#> 19  9.129784  9.107566  8.172159  8.977241  7.615835  9.297364  8.253402
#> 20  8.917501  8.370681  8.858650  9.568423  8.331439  7.929819  9.665785
#> 21  7.205355  7.793396  8.334727  8.616616 10.706891  6.068718  8.692156
#> 22  8.101276  6.722169  6.887710  6.364406  7.684337  8.324979  9.485813
#> 23  9.267573  9.558030  9.736762 10.480360  8.825842 11.004605  8.857251
#> 24  7.626989  7.519196  7.531355  8.281885  9.359565  8.636901 10.468975
#> 25  8.410603  8.734463  6.389538  9.688766  7.331003  9.271481 10.205297
#> 26  7.889876  7.276893  6.461325  8.622568  9.881721  8.732522  9.169574
#> 27 10.601752  7.931032  8.846036  8.985508  8.103005  9.493964  9.210217
#> 28  8.387307  6.755694  8.519988  7.628057  8.734622  9.798813  8.799071
#> 29  8.300193  6.705612  7.475023  9.361034  9.154494        NA        NA
#> 30  8.255262  8.136895  9.247702  9.337644  9.443362  6.654619  9.298538
#> 31  8.447013 10.268032  8.251916  9.503211  7.649532  9.689996  7.752980
#> 32  8.362507  8.178234  8.522558  9.790590  9.628914  7.267687  9.248609
#> 33  9.527252  9.512679  7.804466  7.317895  7.696994  9.434813  8.743502
#> 34  8.542337  9.212701  8.312060  8.231255 11.274211  8.714363  8.453562
#> 35  9.150814  9.228830  6.972037  8.714274  8.930373  8.129145  8.235975
#> 36  8.121059  7.507155  6.718975  7.834465  6.496100  8.225657  7.485531
#> 37  8.189289  8.430200  8.571913  9.286215  7.607443  9.765364  7.599524
#> 38  7.606171  8.492697  8.439719  7.282467  8.323848  9.199037  7.788449
#> 39  8.048048  9.938053 10.267118  9.800409  8.503688  8.502213  9.509453
#> 40  8.395632  6.813602  8.793737  8.227300  8.333663  7.610184  7.809753
#> 41  8.133676  8.595831  7.544810  6.561000  7.562573  9.274109  9.361220
#> 42        NA        NA        NA        NA        NA        NA        NA
#> 43  8.424766  9.494871  9.788712  8.518093  8.776018 10.139155  7.353374
#> 44  7.488138  7.242308  6.716662  8.594243  9.759944  8.992595  7.659990
#> 45  8.861757  9.918389  9.428486  8.245199  9.701237  9.018403  7.894979
#> 46  9.019771  9.126949  7.617221  8.410704  7.602370  9.675589  7.926070
#> 47  6.941166  6.944646  7.885749  8.382798  7.591136  9.807136  8.021478
#> 48  6.649652  9.284329  6.162310  7.674578  5.918886  8.390648  7.308685
#> 49  8.424367  9.211236  7.554145  7.853692  8.024164  8.513923  6.838906
#> 50  8.055747  8.539875  8.459194  8.666338 10.010627  9.047045  8.463236
#> 51  7.549829  9.600416  8.479486  8.670218  8.577561 10.177134  8.534863
#> 52  8.749618  7.589702  9.836609 10.108509  8.519467  7.534553  7.276382
#> 53  8.529054  8.064235  8.065122  8.131127  7.462586  7.083162  7.437610
#> 54  8.905778  6.921674  8.056416  7.636532  9.526859  8.552792  6.624388
#> 55  8.266039  8.418364  8.868945        NA        NA        NA        NA
#> 56  7.300712 10.752262  6.703033  7.401707  7.120570  7.289240  8.435209
#> 57  9.112329  7.599242  7.266966  6.232186  6.932337  9.122584  6.589002
#> 58  9.926244  9.495111  7.885450  8.458552  7.036663  9.255492 10.897277
#> 59  8.664952  8.888850  9.891502  8.850378 11.324596 10.295284  9.982981
#> 60  7.586291  6.835939  9.819137  8.074722  8.428340  9.762091  7.216018
#> 61        NA        NA        NA        NA        NA        NA        NA
#> 62 10.192118  8.135374  8.467315 12.575277  8.480961  8.931632  9.512269
#> 63  8.433294  9.139368 10.477260  8.433941  8.751912  8.508802  7.718103
#> 64  8.210780  8.466184  8.233188  8.415152 10.188713  7.270900  7.684162
#> 65  8.929580  6.766641  6.994610  6.947923  8.114584  8.079705  7.831591
#> 66 10.275864  6.471850  7.442565  6.428799  7.606972  6.564277  7.709019
#> 67  6.478185  7.794688  6.609729  8.430414  8.309080  7.863831  8.680573
#> 68  9.448144  7.571734  7.184959  8.665176  7.312291  8.894767  8.794916
#> 69  8.165244  7.159388  9.332671  9.970292  8.920148 10.188260  9.196200
#> 70  8.555912  8.097944  7.589723  9.023545        NA        NA        NA
#> 71  7.504315  7.024561  9.355983  9.566582  9.620843  8.507943  9.542588
#> 72  8.552512  8.265201  9.369226  7.732812  8.760950  9.719509  8.242263
#> 73  7.976576  8.213302  8.351554  9.062660  7.380792  9.535653  7.366767
#> 74  8.777316  7.235877  9.785779  9.887617  8.329890  8.212853  9.793975
#> 75  8.494482  9.044626  7.652951  9.724337  8.754780  8.638813  7.444064
#> 76  8.832437  7.328571  9.519323  7.827151  8.096762  8.038803  8.450078
#> 77  8.955571  7.343308  9.324979  6.725874  9.472196 10.428724  8.336330
#> 78  7.887072  8.121454  8.464470  8.090119  9.279391  8.519172  8.482158
#> 79  8.063736  8.573971  9.705250  8.909718  8.218468 10.240779  8.163315
#> 80  8.953033 10.059517  9.539249  9.934299  8.358561 10.601384  8.513777
#> 81  8.537614  7.048450  9.563358  8.637722  8.902719  7.896456  9.821129
#> 82  7.211441  8.977665  8.713830  9.222801  9.065650  8.672831  7.619867
#> 83        NA        NA        NA        NA        NA        NA        NA
#> 84  8.104894  8.907404  8.330050  9.513243  7.112694  7.395117  6.742482
#> 85  9.709944  8.776257  9.907470        NA        NA        NA        NA
#> 86  7.844577  8.799663  9.417089  7.913234 10.143819 10.429977        NA
#> 87  9.150383  6.798691  8.473366  8.036621  8.296578  8.855287  7.884383
#> 88  7.018657  7.613447 10.743266  8.890898  9.862574 10.149180  7.994142
#> 89  7.355366  8.173207  7.008531  8.582986  7.695641  8.962534        NA
#> 90        NA        NA        NA        NA        NA        NA        NA
#> 91  9.407727  7.754121  9.780463  9.635164 10.613082 10.412239  9.472230
#> 92  8.013494  7.363688  7.825411  8.203299 10.054059  8.969506  8.636456
#> 93  9.829624  9.248811  9.746113  8.797910  8.141215  8.454898  8.417680
#> 94  9.191738  8.059491  8.072575  7.231174  8.111317  7.777202  9.284333
#> 95  8.374800  7.225673  9.691585  7.386602  9.531471  9.057812  8.384301
#> 96  7.954855  5.827284  8.228933  8.923630  8.508915  7.718435  7.933098
#> 97  8.083419  9.110354  9.675537  7.859437  8.872055  6.714407  8.871208
#> 98  9.382289  8.321610 10.582013  6.743703  8.773795  7.854227  7.514265
#>    any_rescue rescue_start
#> 1        TRUE            7
#> 2       FALSE           NA
#> 3       FALSE           NA
#> 4        TRUE           11
#> 5       FALSE           NA
#> 6       FALSE           NA
#> 7       FALSE           NA
#> 8       FALSE           NA
#> 9       FALSE           NA
#> 10      FALSE           NA
#> 11      FALSE           NA
#> 12      FALSE           NA
#> 13      FALSE           NA
#> 14      FALSE           NA
#> 15      FALSE           NA
#> 16      FALSE           NA
#> 17      FALSE           NA
#> 18       TRUE            2
#> 19       TRUE           11
#> 20      FALSE           NA
#> 21      FALSE           NA
#> 22      FALSE           NA
#> 23       TRUE            3
#> 24      FALSE           NA
#> 25      FALSE           NA
#> 26      FALSE           NA
#> 27      FALSE           NA
#> 28      FALSE           NA
#> 29      FALSE           NA
#> 30      FALSE           NA
#> 31      FALSE           NA
#> 32      FALSE           NA
#> 33      FALSE           NA
#> 34      FALSE           NA
#> 35      FALSE           NA
#> 36      FALSE           NA
#> 37       TRUE            2
#> 38       TRUE           11
#> 39      FALSE           NA
#> 40      FALSE           NA
#> 41      FALSE           NA
#> 42       TRUE            7
#> 43       TRUE           11
#> 44      FALSE           NA
#> 45      FALSE           NA
#> 46       TRUE           11
#> 47      FALSE           NA
#> 48       TRUE           11
#> 49      FALSE           NA
#> 50      FALSE           NA
#> 51      FALSE           NA
#> 52      FALSE           NA
#> 53      FALSE           NA
#> 54      FALSE           NA
#> 55      FALSE           NA
#> 56      FALSE           NA
#> 57      FALSE           NA
#> 58      FALSE           NA
#> 59      FALSE           NA
#> 60      FALSE           NA
#> 61      FALSE           NA
#> 62      FALSE           NA
#> 63      FALSE           NA
#> 64      FALSE           NA
#> 65      FALSE           NA
#> 66      FALSE           NA
#> 67      FALSE           NA
#> 68      FALSE           NA
#> 69      FALSE           NA
#> 70      FALSE           NA
#> 71      FALSE           NA
#> 72      FALSE           NA
#> 73       TRUE            1
#> 74      FALSE           NA
#> 75      FALSE           NA
#> 76      FALSE           NA
#> 77       TRUE           10
#> 78      FALSE           NA
#> 79      FALSE           NA
#> 80       TRUE            8
#> 81      FALSE           NA
#> 82      FALSE           NA
#> 83      FALSE           NA
#> 84      FALSE           NA
#> 85       TRUE            8
#> 86       TRUE           10
#> 87      FALSE           NA
#> 88       TRUE           10
#> 89      FALSE           NA
#> 90      FALSE           NA
#> 91      FALSE           NA
#> 92      FALSE           NA
#> 93      FALSE           NA
#> 94      FALSE           NA
#> 95      FALSE           NA
#> 96      FALSE           NA
#> 97      FALSE           NA
#> 98      FALSE           NA
Design <- assumptions_diabetes_rescue()
Design
#>   delta    lambda delta_resc lambda_resc  k mean_bl mean_age sd_age      b_age
#> 1     1 0.3465736       0.75           1 12       8       60     10 0.06931472
#>   pr_rescue      h_y        h_age pr_missing       g_y      g_age  g_rescue
#> 1      0.05 1.098612 -0.009950331       0.02 0.4054651 0.01980263 0.4054651
true_summary_statistics_diabetes_rescue(assumptions_diabetes_rescue())
#>   rmst_trt medial_surv_trt rmst_ctrl median_surv_ctrl gAHR AHR
#> 1       NA              NA        NA               NA   NA  NA